requirements.txt
opencv-python
pytesseract
numpy
Pillow
flask 